remove restriction of 0 params for compiling with clrjit
